corrijo

+++ src/components/Turnos/CardTurno.jsx
@@ -13,34 +13,58 @@ import {
 import { useDispatch, useSelector } from "react-redux";
 
 import CloseButton from "react-bootstrap/CloseButton";
-import { deleteTurno } from "../axios/axios.turnos";
+import { deleteTurno, getUserTurnos } from "../axios/axios.turnos";
 import ListGroup from "react-bootstrap/ListGroup";
 import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
-import Container from 'react-bootstrap/Container';
-import Row from 'react-bootstrap/Row';
-import Col from 'react-bootstrap/Col';
-import Stack from 'react-bootstrap/Stack';
+import Container from "react-bootstrap/Container";
+import Row from "react-bootstrap/Row";
+import Col from "react-bootstrap/Col";
+import Stack from "react-bootstrap/Stack";
 
-const CardTurno = ({ date, hour, nombre, surname, status, _id }) => {
+const CardTurno = ({ date, hour, status, _id }) => {
   const dispatch = useDispatch();
   const { currentUser } = useSelector((state) => state.user);
 
   const handleClick = async (e) => {
-    const turno = {
-      _id: _id,
-    };
-    await deleteTurno(currentUser.token, turno);
+    const confirmar = window.confirm("Desea eliminar el turno?");
+    if (confirmar) {
+      const turno = {
+        _id: _id,
+      };
+      await deleteTurno(currentUser.token, turno);
+      await getUserTurnos(currentUser.token, dispatch);
+    }
   };
   return (
-    <Container fluid="lg">
+    <div className="card col-lg-6  ">
+      <div className="card-header d-flex gap-5">
+        <div>Turno: #{_id.slice(16)}</div>
+        <div className="ms-auto">
+          <CloseButton onClick={(e) => handleClick(e)} />
+        </div>
+      </div>
+      <div className="card-body">
+        <ul className="list-group list-group-flush">
+          <li className="list-group-item">Dia: {date}</li>
+          <li className="list-group-item">Hora: {hour}</li>
+          <li className="list-group-item">Estado: {status}</li>
+        </ul>
+      </div>
+    </div>
+  );
+};
+
+export default CardTurno;
+
+{
+  /* <Container fluid="lg">
       <Row>
-        
-        <Col  xs={12} md={8}>
+        <Col  >
           <Card>
             <Card.Header >
             <Stack direction="horizontal" gap={3}>
-              <div>Turno: {_id}</div>
+              <div>Turno: #{_id.slice(16)}</div>
               <div className="ms-auto"><CloseButton onClick={(e) => handleClick(e)}/></div>
               </Stack>
             </Card.Header>
@@ -53,13 +77,9 @@ const CardTurno = ({ date, hour, nombre, surname, status, _id }) => {
             </Card.Body>
           </Card>
         </Col>
-        
       </Row>
-    </Container>
-  );
-};
-
-export default CardTurno;
+    </Container> */
+}
 
 {
   /* <TurnoContainerStyled>
